Add documentation with Valadoc
This uses Valadoc to generate the GIR file instead of the Vala compiler;
because Valadoc can insert documentation comments into the GIR file.

These are not correctly formatted Valadoc comments, and that is because
I'm planning to build the documentation with g-ir-doctool, not Valadoc;
because the library should have documentation for more languages than
just Vala.