Port "wait_for" mechanisms from Valadate
Valadate is not maintained anymore, but it has a few things which would be
quite useful in this library: wait_for_signal() and wait_for_async(). This
ports those mechanisms over, pretty much wholesale. We include the tests
that were present in Valadate, but those are now written in C in order to
make sure that the API is still reasonable to use from C.

One thing had to change; Valadate disconnected the safeguard timers even
after they had timed out. That causes a warning to be logged, which would
make the normal GLib test runner fail a test.

There is still one bug to fix: we must push and pop thread-default main
contexts when using the wait functions so that a callback from one test's
main loop doesn't pollute that of another test.

NOTE: The Valadate code was licensed under LGPLv3, but per [1], it is
allowed to relicense a derivative work of it under GPLv3 without
the express permission of the original authors.

[1] http://www.gnu.org/licenses/gpl-faq.html#AllCompatibility