WIP custom contexts
This should solve the big FIXME in the Valadate code about pushing and
popping custom main loop contexts during tests so that callbacks that are
scheduled in one main loop don't pollute the main loop in a subsequent
test.

One unanswered question is why we get the following warning

src/wait.vala:62.25-62.32: warning: copying delegates is not supported

when adding a timeout to a custom main loop context.

The other is whether it would be better to push and pop the custom main
loop contexts in the test setup and teardown functions. We could of
course provide convenience API for this. That would allow the custom
contexts to be optional in case external code didn't follow good
practices around GMainContext (see [1]).

Pushing and popping inside the wait functions themselves also makes it
impossible to do any setup that queues events before entering the wait
function; you have to use the provided GtBlock callback.

[1] https://tecnocode.co.uk/2014/03/27/what-is-gmaincontext/