Delegate to the local VFS when appropriate
I misunderstood how to implement a GVfs, and thought that GtMockVfs'
methods would only be called if it had been determined that a mock file
was wanted.

Instead, you have to delegate to the local VFS if your URI scheme does not
match the ones you handle in your custom VFS; and you always have to
delegate in get_file_for_path() because otherwise you can't get local
files by their pathname anymore. (I thought that get_file_for_path() just
wouldn't be called in GtMockVfs, but it is.)